Reorganized DataSourceProcessor tests.  One was in the wrong class and another had the wrong name.
