+++ core/src/test/java/org/mapfish/print/processor/jasper/MergeDataSourceProcessorIntegrationTest.java
@@ -32,7 +32,6 @@ import org.mapfish.print.processor.ProcessorDependency;
 import org.mapfish.print.processor.ProcessorDependencyGraph;
 import org.mapfish.print.processor.ProcessorGraphNode;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.annotation.DirtiesContext;
 
 import java.io.File;
 import java.util.List;
@@ -41,7 +40,6 @@ import java.util.regex.Pattern;
 import javax.annotation.Nonnull;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
 
 public class MergeDataSourceProcessorIntegrationTest extends AbstractMapfishSpringTest {
 
@@ -54,17 +52,6 @@ public class MergeDataSourceProcessorIntegrationTest extends AbstractMapfishSpri
     };
     @Autowired
     private ConfigurationFactory configurationFactory;
-
-    @Test @DirtiesContext
-    public void testValidate() throws Exception {
-        final File configFile = getFile("incorrectly-configured-DataSourceProcessor/config.yaml");
-        configurationFactory.setDoValidation(false);
-        final Configuration config = configurationFactory.getConfig(configFile);
-
-        final List<Throwable> validate = config.validate();
-
-        assertTrue(validate.size() > 0);
-    }
 
     @Test
     public void testCreateDependencies() throws Exception {
